refactor: 1주차 코드 컴포넌트화
