+++ src/Components/Section/SelectFriend.tsx
@@ -1,59 +1,35 @@
-/** @jsxImportSource @emotion/react */
-import { css, useTheme } from '@emotion/react';
-import type { ThemeType } from '@/styles/theme';
-import { FaPlus } from 'react-icons/fa';
+import styled from '@emotion/styled';
 
-const wrapper = css`
+const Wrapper = styled.div`
   padding: 16px;
-  background-color: #f5f6f8;
+  background-color: ${({ theme }) => theme.color.semantic.background.default};
 `;
 
-const container =css `
+const Container = styled.div`
   display: flex;
   align-items: center;
-  background-color: white;
   border-radius: 20px;
-  padding: 20px;
-  box-shadow: 0 1px 3px rgba(0, 0, 0, 0.05);
+  padding: 30px;
+  background-color: ${({ theme }) => theme.color.yellow.yellow600};
 `;
 
-const iconWrapper = (theme: ThemeType) => css`
-  width: 48px;
-  height: 48px;
-  background-color: ${theme.color.yellow.yellow600};
-  border-radius: 50%;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  margin-right: 16px;
+const Text = styled.div`
+  font-size: 1rem;
+  font-weight: 700;
+  line-height: 1.1875rem;
+  color: rgb(42, 48, 56);
+  margin: 0;
+  text-align: left;
 `;
 
-const text = (theme: ThemeType) => css`
-  color: ${theme.color.semantic.text.default};
-  font-size: 16px;
-  font-weight: 500;
-
-  a {
-    text-decoration: underline;
-    color: ${theme.color.blue.blue700};
-  }
-`;
-
-const SelectFriend = () => {
-  const theme = useTheme() as ThemeType;
-
+const Banner = () => {
   return (
-    <div css={wrapper}>
-      <div css={container}>
-        <div css={iconWrapper(theme)}>
-          <FaPlus color="black" size={16} />
-        </div>
-        <div css={text(theme)}>
-          선물할 친구를 선택해 주세요.
-        </div>
-      </div>
-    </div>
+    <Wrapper>
+      <Container>
+        <Text>프론트엔드 2단계 과제 화이팅!</Text>
+      </Container>
+    </Wrapper>
   );
 };
 
-export default SelectFriend;
+export default Banner;
